Update overview.tpl
This updated multiple issues with the client screen. This adds a button to open the Instance ID in a new tab, It shows the Endpoints in a URL format and allows it to be copied to clipboard, new wording has been provided to ensure customers Start the Instance before trying to access it.